Fundir as entidades User e Seller
As entidades User e Seller foram fundidas em um só: a entidade
User. A diferenciação entre os dois tipos será através do campo
"isSeller" e pelo controle da camada de negócios.
