+++ src/main/java/br/com/bootcamp/desafio_spring/entity/User.java
@@ -1,8 +1,12 @@
 package br.com.bootcamp.desafio_spring.entity;
 
+import java.util.List;
+
 public class User implements IEntity {
     private Integer ID;
     private String name;
+    private boolean isSeller;
+    private List<Post> posts;
 
     public User() {
     }
@@ -10,6 +14,12 @@ public class User implements IEntity {
     public User(Integer ID, String name) {
         this.ID = ID;
         this.name = name;
+    }
+
+    public User(Integer ID, String name, boolean isSeller) {
+        this.ID = ID;
+        this.name = name;
+        this.isSeller = isSeller;
     }
 
     @Override
@@ -28,6 +38,22 @@ public class User implements IEntity {
 
     public void setName(String name) {
         this.name = name;
+    }
+
+    public boolean isSeller() {
+        return isSeller;
+    }
+
+    public void setSeller(boolean seller) {
+        isSeller = seller;
+    }
+
+    public List<Post> getPosts() {
+        return posts;
+    }
+
+    public void setPosts(List<Post> posts) {
+        this.posts = posts;
     }
 
     @Override
